ref: Extracts the VideoPolicy from the VideoConstraints.
This commit separates the _video policy_ from the _video constraints_.

Video constraints express the physicall constraints of a video source.
These are used from the bridge in both its video sending and receiving
function: When the bridge acts as a sender they're used to limit the
choice of layers to consider when performing bandwidth distribution.
When the bridge acts as a receiver, it feeds these constraints back to
the client so that the client can reduce its bitrate.

The video policy defines preferences and carries instructions on how to
perform bandwdith distribution, so they are irrelevant in the video
receiving function of the bridge. These instructions are internally
translated as additional constraints by the bandwidth distribution algo.

The motivation behind this commit is to clean-up the `VideoConstraints`
class from information that is irrelevant for the receive-side aspect of
the bridge.